requirements.txt
asgiref==3.7.2
cmake==3.29.3
Django==4.2.10
django-rest-framework==0.1.0
djangorestframework==3.15.1
numpy==1.26.4
opencv-python==4.9.0.80
sqlparse==0.4.4
typing_extensions==4.9.0
